Update maybe to be agnostic about success or error
A Maybe has a value or not. We don't want to advertise whether we think that value is a 'success' or an 'error'
Add a Result<T1, T2>.Invert() method to swap T1 and T2. This is useful for when Maybe<TErr>.ToResult() is called and we want to change it to an error result.