ICanBeValid interface and extension methods
Some objects and especially structs need to be able to tell when they are in a valid state.
A simple .IsValid property can work wonders there.
Add a few extension methods so we can work with self-validatable objects with an interface similar to Result<,> and Maybe<>
Also methods to convert to Result<,> and Maybe<>
We set these methods up as generics with generic receivers, to avoid boxing struct types