Cursor and CursorPage types
The Paging/Page types typically want to turn into a database query with 'LIMIT @pageSize OFFSET @startOffset' this is fine for small data sets, but for large data sets this can put a lot of strain on the DB
Cursor/CursorPage instead uses some kind of id/key to start with. This turns into a DB query like 'WHERE id > @lastKey' which is faster on the db because we don't need to compile and order the entire dataset just to take a small chunk out of it. We instead jump directly to the start of the next result set and continue our iteration.
Cursor is better for infinite scroll scenarios and not as good for pagination scenarios where we need to jump directly to Page 5 or whatever