Implementation of the Either monad
Borrowed a lot of code from Result, but had to remove several methods which did not make sense
Result is a specialization of Either, with the added assumption that the right value is an error. This allows us to do things like shortcut certain operations when we're in an error state. .Bind/.And makes a lot more sense in the context of Result than it does in the context of Either, for example
Without this added heuristic, Either has less raw functionality but it is still an efficient way to return one of two values from an operation