Rework specifications
IQuerySpecification -> ISpecification for brevity
Add ISpecification.IsSatisfiedBy() method
Add an ExpressionSpecification for delegates/expressions
Cleanup combine logic and move it into the CombinedSpecification. We want to make some guarantees that the list is immutable and never contains a fail or a permissive
Make several constructors private. Prefer singleton instances for fail and permissive. Prefer a factory method for Combine